% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usensys_maps}
\alias{usensys_maps}
\title{GIS/map-objects for alternative versions of the model.}
\format{
A list of sf-objects:
\describe{
\item{us26_sf}{26-region map}
}
}
\usage{
usensys_maps
}
\description{
GIS/map-objects for alternative versions of the model.
}
\examples{
# Plot with base R
plot(usensys_maps$us26_sf["region"])

# Plot with ggplot2
library(ggplot2)
ggplot(usensys_maps$us26_sf) +
geom_sf(aes(fill = region), show.legend = F) +
# geom_point(aes(x = centr_lon, y = centr_lat), color = "red", shape = 20) +
geom_label(aes(x = centr_lon, y = centr_lat, label = region), size = 3) +
scale_fill_viridis_d(option = "H") +
labs(x = "", y = "") +
theme_void()
}
\keyword{datasets}
