% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usensys_maps}
\alias{usensys_maps}
\title{GIS/map-objects for alternative versions of the model.}
\format{
A list of sf-objects:
\describe{
\item{us26_sf}{26-region map}
}
}
\usage{
usensys_maps
}
\description{
GIS/map-objects for alternative versions of the model.
}
\examples{
plot(usensys_maps$us26_sf["region"])
library(ggplot2)
ggplot() + geom_sf(data = usensys_maps$us26_sf, aes(fill = region)) + theme_bw()
}
\keyword{datasets}
